Moved most of the code to CheckDataGrabber
