+++ src/js/components/core/SwitchCheckModal.js
@@ -3,18 +3,14 @@ const React = require('react');
 const path = require('path');
 const Button = require('react-bootstrap/lib/Button.js');
 const Modal = require('react-bootstrap/lib/Modal.js');
-const CheckStore = require('../../stores/CheckStore.js');
 const CoreStore = require('../../stores/CoreStore.js');
 const CoreActions = require('../../actions/CoreActions.js');
 const FileModule = require('./FileModule.js');
 const CheckDataGrabber = require('./create_project/CheckDataGrabber.js');
 
-var _this;
-
 class SwitchCheckModal extends React.Component{
   constructor(){
     super();
-    _this = this
     this.state ={
       showModal: false,
     }
@@ -37,22 +33,8 @@ class SwitchCheckModal extends React.Component{
   }
 
   moduleClick(folderName) {
-    FileModule.readJsonFile(path.join(window.__base, "modules", folderName, "manifest.json"), (jsonObject) => {
-      // hard coded
-      var params = {originalLanguagePath: "C:/Users/Logan Lebanoff/Desktop/8woc/8woc/data/ulgb", targetLanguagePath: "C:/Users/Logan Lebanoff/Desktop/8woc/test_files/Ephesians", repo: undefined, bookAbbr: "eph"}
-      var array = [];
-      array.push([jsonObject.name, path.join(window.__base, "modules", folderName)]);
-      for(let childFolderName of jsonObject.include) {
-        array.push([childFolderName, path.join(window.__base, "modules", childFolderName)]);
-      }
-      // hard coded
-      if(!CheckStore.storeData.common) {
-        CheckStore.storeData.common = {};
-        CheckStore.storeData.common.saveLocation = "C:/Users/Logan Lebanoff/Desktop/8woc/test_files/Ephesians";
-      }
-      debugger;
-      CheckDataGrabber.getFetchData(array, params);
-    });
+    this.close();
+    CheckDataGrabber.loadModuleAndDependencies(folderName);
   }
 
   updateCheckModal() {
